Add Cards Content Area Controls and move Cards Collection controls in separate file
