+++ src/blocks/cards-collection/scss/_style.scss
@@ -8,3 +8,16 @@
 .novablocks-cards-collection__layout {
 	@include novablocks-cards-collection-layout;
 }
+
+@each $alignment in left, center, right {
+
+	.novablocks-cards-collection--align-#{ $alignment } {
+		text-align: $alignment;
+	}
+}
+
+.novablocks-cards-collection--align-right {
+	.wp-block-buttons {
+		flex-direction: row-reverse;
+	}
+}
